+++ libnd4j/include/ops/declarable/generic/list/pick_list.cpp
@@ -30,7 +30,7 @@ namespace ops {
 LIST_OP_IMPL(pick_list, 1, 1, 0, -2) {
   auto list = INPUT_LIST(0);
 
-  std::vector<int> indices;
+  std::vector<LongType> indices;
   if (block.width() > 1 && block.getVariable(1)->getNDArray()->isVector()) {
     auto ia = INPUT_VARIABLE(1);
     for (int e = 0; e < ia->lengthOf(); e++) indices.emplace_back(ia->e<int>(e));
