+++ libnd4j/include/ops/declarable/helpers/percentile.h
@@ -31,7 +31,7 @@ namespace ops {
 namespace helpers {
 
 SD_LIB_HIDDEN void percentile(sd::LaunchContext* context, const NDArray& input, NDArray& output,
-                              std::vector<int>& axises, const float q, const int interpolation);
+                              std::vector<LongType>& axises, const float q, const int interpolation);
 
 }
 }  // namespace ops
